project/requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3


#pip install django     (MAIN)
#pip install djangorestframework    (REST)
#pip install python-dotenv  (ENV)